+++ examples/django/charm/src/charm.py
@@ -10,12 +10,12 @@ import typing
 
 import ops
 
-import paas_app_charmer.django
+import paas_charm.django
 
 logger = logging.getLogger(__name__)
 
 
-class DjangoCharm(paas_app_charmer.django.Charm):
+class DjangoCharm(paas_charm.django.Charm):
     """Django Charm service."""
 
     def __init__(self, *args: typing.Any) -> None:
