+++ paas_charm/app.py
@@ -12,8 +12,8 @@ from typing import List
 
 import ops
 
-from paas_app_charmer.charm_state import CharmState, IntegrationsState
-from paas_app_charmer.database_migration import DatabaseMigration
+from paas_charm.charm_state import CharmState, IntegrationsState
+from paas_charm.database_migration import DatabaseMigration
 
 logger = logging.getLogger(__name__)
