+++ paas_charm/django/cos/prometheus_alert_rules/django.rule
