+++ paas_charm/fastapi/__init__.py
@@ -1,0 +1,8 @@
+# Copyright 2024 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""FastAPI module entrypoint."""
+
+from .charm import Charm
+
+__all__ = ["Charm"]
