+++ pyproject.toml
@@ -1,7 +1,7 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 [project]
-name = "paas-app-charmer"
+name = "paas-charm"
 version = "1.3.1"
 description = "Companion library for PaaS app charmer."
 readme = "README.md"
@@ -20,13 +20,13 @@ classifiers = [
 
 
 [project.urls]
-Repository = "https://github.com/canonical/paas-app-charmer"
+Repository = "https://github.com/canonical/paas-charm"
 
 [tool.setuptools.dynamic]
 dependencies = {file = ["requirements.txt"]}
 
 [tool.setuptools.package-data]
-paas_app_charmer = ["**/cos/**", "**/cos/**/.**"]
+paas_charm = ["**/cos/**", "**/cos/**/.**"]
 
 # Testing tools configuration
 [tool.coverage.run]
