+++ tests/integration/flask/conftest.py
@@ -54,7 +54,7 @@ async def charm_file_fixture(pytestconfig: pytest.Config, ops_test: OpsTest) -> 
         charm_file = await ops_test.build_charm(PROJECT_ROOT / "examples/flask")
     elif charm_file[0] != "/":
         charm_file = PROJECT_ROOT / charm_file
-    inject_venv(charm_file, PROJECT_ROOT / "paas_app_charmer")
+    inject_venv(charm_file, PROJECT_ROOT / "paas_charm")
     return pathlib.Path(charm_file).absolute()
