+++ tests/unit/django/conftest.py
@@ -14,7 +14,7 @@ import pytest
 from ops.testing import Harness
 
 from examples.django.charm.src.charm import DjangoCharm
-from paas_app_charmer.database_migration import DatabaseMigrationStatus
+from paas_charm.database_migration import DatabaseMigrationStatus
 
 from .constants import DEFAULT_LAYER
