+++ tests/unit/flask/test_charm.py
@@ -14,12 +14,12 @@ import pytest
 from ops.pebble import ServiceStatus
 from ops.testing import Harness
 
-from paas_app_charmer._gunicorn.webserver import GunicornWebserver, WebserverConfig
-from paas_app_charmer._gunicorn.workload_config import create_workload_config
-from paas_app_charmer._gunicorn.wsgi_app import WsgiApp
-from paas_app_charmer.charm_state import CharmState
-from paas_app_charmer.database_migration import DatabaseMigrationStatus
-from paas_app_charmer.flask import Charm
+from paas_charm._gunicorn.webserver import GunicornWebserver, WebserverConfig
+from paas_charm._gunicorn.workload_config import create_workload_config
+from paas_charm._gunicorn.wsgi_app import WsgiApp
+from paas_charm.charm_state import CharmState
+from paas_charm.database_migration import DatabaseMigrationStatus
+from paas_charm.flask import Charm
 
 from .constants import (
     DEFAULT_LAYER,
