+++ tests/unit/flask/test_flask_app.py
@@ -12,11 +12,11 @@ import unittest.mock
 
 import pytest
 
-from paas_app_charmer._gunicorn.webserver import WebserverConfig
-from paas_app_charmer._gunicorn.workload_config import create_workload_config
-from paas_app_charmer._gunicorn.wsgi_app import WsgiApp
-from paas_app_charmer.app import map_integrations_to_env
-from paas_app_charmer.charm_state import CharmState, IntegrationsState, S3Parameters
+from paas_charm._gunicorn.webserver import WebserverConfig
+from paas_charm._gunicorn.workload_config import create_workload_config
+from paas_charm._gunicorn.wsgi_app import WsgiApp
+from paas_charm.app import map_integrations_to_env
+from paas_charm.charm_state import CharmState, IntegrationsState, S3Parameters
 
 
 @pytest.mark.parametrize(
