+++ tests/unit/go/test_app.py
@@ -8,9 +8,9 @@ import unittest
 
 import pytest
 
-from paas_app_charmer.app import App, WorkloadConfig
-from paas_app_charmer.charm_state import CharmState, IntegrationsState
-from paas_app_charmer.go.charm import GoConfig
+from paas_charm.app import App, WorkloadConfig
+from paas_charm.charm_state import CharmState, IntegrationsState
+from paas_charm.go.charm import GoConfig
 
 
 @pytest.mark.parametrize(
