+++ tox.ini
@@ -7,10 +7,11 @@ envlist = lint, unit, static, coverage-report
 skip_missing_interpreters = True
 
 [vars]
-src_path = {toxinidir}/paas_app_charmer/
+src_path = {toxinidir}/paas_charm/
+legacy_src_path ={toxinidir}/paas_app_charmer/
 example_path = {toxinidir}/examples/
 tst_path = {toxinidir}/tests/
-all_path = {[vars]src_path} {[vars]example_path} {[vars]tst_path}
+all_path = {[vars]src_path} {[vars]example_path} {[vars]tst_path} {[vars]legacy_src_path}
 
 [testenv]
 setenv =
@@ -67,12 +68,12 @@ commands =
       --skip {toxinidir}/examples/flask/lib \
       --skip {toxinidir}/tests/integration/.mypy_cache
     # pflake8 wrapper supports config from pyproject.toml
-    pflake8 {[vars]src_path}
+    pflake8 {[vars]src_path} {[vars]legacy_src_path}
     isort --check-only --diff {[vars]all_path}
     black --target-version=py310 --check --diff {[vars]all_path}
-    mypy {[vars]src_path}
-    pydocstyle {[vars]src_path}
-    pylint {[vars]src_path} --min-similarity-lines=10
+    mypy {[vars]src_path} {[vars]legacy_src_path}
+    pydocstyle {[vars]src_path} {[vars]legacy_src_path}
+    pylint {[vars]src_path} {[vars]legacy_src_path} --min-similarity-lines=10
 
 [testenv:static]
 description = Run static analysis tests
@@ -90,7 +91,7 @@ deps =
     coverage[toml]
     -r{toxinidir}/requirements.txt
 commands =
-    coverage run --source={[vars]src_path} \
+    coverage run --source={[vars]src_path},{[vars]legacy_src_path} \
         -m pytest --ignore={[vars]tst_path}integration -v --tb native -s {posargs}
     coverage report
